Attempt to account for record coercion
When a record is coerced within Log::write(), the record passed to
the stream and filter is different/new than the one passed to
Log::write(). For now track the original one, but keep the coerced
version for later writing. This isn't quite right though as a
Log::delay() within a filter or policy hook won't do the right
thing in these scenarios.