Dabble with a log write batch event
Instead of transporting writes in a custom broker message, what if we
packed it into a custom event.

    event Log::write_batch(stream: ID, writer: Writer, path: string, format: BatchFormat, n: count, payload: string)

The payload would be a custom encoding of individual log writes using
whatever format BatchFormat indicates. This implementation uses the
existing BinarySerializationFormat.

The event handler could be provided as C++ by the logging Manager.