cluster/zeromq: POC cluster backend using XPUB/XSUB manager zmq::proxy()
This is nowhere near complete and a bit unsure about threading, but
enough to see events being published and test a bit. Main quirk is that
a worker publishing to the worker_topic will see its own messages.
It's something that NATS does by default, too. Probably need to include
our own nodeid into the message so we can ignore it upon receipt.

Otherwise, this isn't all that bad.