logging: Dedicated log flush timer
Log flushing is currently triggered via the threading heartbeat timer.
The writer backend enqueues a message for the frontend, which in turn
sends a write message and flush message down to the backend again.

This change introduces a separate timer that is managed by the logger
manager instead of piggy-backing on the heartbeat timer.

This allows to modify the log flush frequency independently of the heartbeat
interval and will later allow the buffering and flushing logic to be reused
for remote cluster backends, too.