script_opt: Do not include ZBody.h in ZAM/Validate.cc
Touching broker/Manager.h or logging/Manager.h or anything that's
included in ZBody.h results in a rebuild of a bunch of files in ZAM.
Particularly, ZAM/Validate.cc takes much too long to recompile for
me (45+ seconds) and apparently that's mostly by accident.

Clean up the ZAM/Validate.cc includes to prevent this.

I'd love to remove some of the includes from ZBody.h because some seem
only needed for ZBody.cc, but due to ZBody.h being included by Compile.h
that requires a rebuild of a lot of files.

ZBody.cc is also annoyingly slow to compile, but to break that include
cycle we'd need need wrapper/trampolines for whatever methods are used
on log_mgr and broker_mgr to break the include dependency.