logging: Dedicated log flush timer
Log flushing is currently triggered via the threading heartbeat timer
and the hard-coded WRITE_BUFFER_SIZE 1000.

This change introduces a separate timer that is managed by the logger
manager instead of piggy-backing on the heartbeat timer.

This allows to modify the log flush frequency and batch size independently
of the threading heartbeat interval. Later, this will allow to re-use the
buffering and flushing logic in the writer frontend for non-Broker cluster
backends, too.

One change here is that even frontends that do not have a backend will
be flushed. This is wanted for non-Broker backends and should be very
cheap. Possibly, Broker can piggy back on this timer down the road, too,
rather than using its own script-level timer (see Broker::log_flush()).